Move styles to Modal component
